requirements.txt
anthropic==0.84.0
langchain-anthropic==1.3.4
langchain-core==1.2.17
langchain-community==0.4.1
langgraph==1.0.10
langgraph-prebuilt==1.0.8
langsmith==0.7.12
pydantic==2.12.5
python-dotenv==1.2.2
duckduckgo-search>=6.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
